added support for float value timeout in CardRequest, thanks to Henryk Plotz

+++ trunk/pyscard/src/smartcard/pcsc/PCSCCardRequest.py
@@ -64,7 +64,7 @@ class PCSCCardRequest(AbstractCardRequest):
             self.timeout=INFINITE
         # otherwise, from seconds to milliseconds
         else:
-            self.timeout=1000*self.timeout
+            self.timeout=int(1000*self.timeout)
